Component 방식으로 수정 (유니티 스타일)

+++ Engine/TableDescriptorHeap.cpp
@@ -8,7 +8,6 @@ void TableDescriptorHeap::Init(uint32 count)
 
 	D3D12_DESCRIPTOR_HEAP_DESC desc = {};
 	desc.NumDescriptors = count * REGISTER_COUNT;
-	// D3D12_DESCRIPTOR_HEAP_FLAG_SHADER_VISIBLE 로 만들어주어야 GPU에 상주가 가능하고 register에 올릴 수 있다.
 	desc.Flags = D3D12_DESCRIPTOR_HEAP_FLAG_SHADER_VISIBLE;
 	desc.Type = D3D12_DESCRIPTOR_HEAP_TYPE_CBV_SRV_UAV;
 
@@ -43,7 +42,6 @@ void TableDescriptorHeap::SetSRV(D3D12_CPU_DESCRIPTOR_HANDLE srcHandle, SRV_REGI
 
 void TableDescriptorHeap::CommitTable()
 {
-	// Register에다가 올려 보내는 부분
 	D3D12_GPU_DESCRIPTOR_HANDLE handle = _descHeap->GetGPUDescriptorHandleForHeapStart();
 	handle.ptr += _currentGroupIndex * _groupSize;
 	CMD_LIST->SetGraphicsRootDescriptorTable(0, handle);
@@ -53,12 +51,12 @@ void TableDescriptorHeap::CommitTable()
 
 D3D12_CPU_DESCRIPTOR_HANDLE TableDescriptorHeap::GetCPUHandle(CBV_REGISTER reg)
 {
-	return GetCPUHandle(static_cast<uint32>(reg));
+	return GetCPUHandle(static_cast<uint8>(reg));
 }
 
 D3D12_CPU_DESCRIPTOR_HANDLE TableDescriptorHeap::GetCPUHandle(SRV_REGISTER reg)
 {
-	return GetCPUHandle(static_cast<uint32>(reg));
+	return GetCPUHandle(static_cast<uint8>(reg));
 }
 
 D3D12_CPU_DESCRIPTOR_HANDLE TableDescriptorHeap::GetCPUHandle(uint8 reg)
@@ -68,4 +66,3 @@ D3D12_CPU_DESCRIPTOR_HANDLE TableDescriptorHeap::GetCPUHandle(uint8 reg)
 	handle.ptr += reg * _handleSize;
 	return handle;
 }
-
